backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
google-cloud-vision==3.7.1
anthropic>=0.34.0
python-docx==1.1.0
pydantic<2.0
python-dotenv==1.0.1
